# fix: reduce loop